[renderer][common] fixed the bug AssistantMessageNode およびこれを含むEdgesが表示されないことがある #6
